requirements.txt
torch
torchvision
pillow
numpy
pandas
scikit-learn
mlflow
fastapi
uvicorn
python-multipart
matplotlib
opencv-python
pydantic
pytest
loguru
